Remove beforeLint feature
This was an underused feature and adds complexity to the preprocessor.
This could be solved more elegantly by allowing the user to pass their
own `transform` function to postcss-import.

Fixes #35

Improve variable naming

This makes it a little clearer what `merged` is

Remove superfluous comments

These are too basic to need comments